Corrected tetengo::gui::win32::widget.
git-svn-id: file:///home/kaoru/svnrepos/bobura/trunk@368 5a212d1d-074f-4316-8284-043cf949738a

+++ tetengo2/include/tetengo2.gui.win32.widget.h
@@ -224,10 +224,10 @@ namespace tetengo2 { namespace gui { namespace win32
 
             const ::BOOL result = ::MoveWindow(
                 this->handle(),
-                position.first,
-                position.second,
-                dimension.first,
-                dimension.second,
+                static_cast<int>(position.first),
+                static_cast<int>(position.second),
+                static_cast<int>(dimension.first),
+                static_cast<int>(dimension.second),
                 this->visible() ? TRUE : FALSE
             );
             if (result == 0)
@@ -266,10 +266,10 @@ namespace tetengo2 { namespace gui { namespace win32
 
             const ::BOOL result = ::MoveWindow(
                 this->handle(),
-                position.first,
-                position.second,
-                dimension.first,
-                dimension.second,
+                static_cast<int>(position.first),
+                static_cast<int>(position.second),
+                static_cast<int>(dimension.first),
+                static_cast<int>(dimension.second),
                 this->visible() ? TRUE : FALSE
             );
             if (result == 0)
